Revert "Use palantir-java-format as default formatter in RemoveUnusedImportsStep"
This reverts commit 2ebe138c
