+++ lib/src/main/java/com/diffplug/spotless/java/RemoveUnusedImportsStep.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2025 DiffPlug
+ * Copyright 2016-2024 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,12 +22,12 @@ import java.util.Objects;
 import com.diffplug.spotless.FormatterStep;
 import com.diffplug.spotless.Provisioner;
 
-/** Uses palantir-java-format or cleanthat.UnnecessaryImport, but only to remove unused imports. */
+/** Uses google-java-format or cleanthat.UnnecessaryImport, but only to remove unused imports. */
 public class RemoveUnusedImportsStep implements Serializable {
 	private static final long serialVersionUID = 1L;
 	static final String NAME = "removeUnusedImports";
 
-	static final String DEFAULT_FORMATTER = "palantir-java-format";
+	static final String GJF = "google-java-format";
 	static final String CLEANTHAT = "cleanthat-javaparser-unnecessaryimport";
 
 	// https://github.com/solven-eu/cleanthat/blob/master/java/src/main/java/eu/solven/cleanthat/engine/java/refactorer/mutators/UnnecessaryImport.java
@@ -37,17 +37,18 @@ public class RemoveUnusedImportsStep implements Serializable {
 	private RemoveUnusedImportsStep() {}
 
 	public static String defaultFormatter() {
-		return DEFAULT_FORMATTER;
+		return GJF;
 	}
 
 	public static FormatterStep create(Provisioner provisioner) {
-		return create(DEFAULT_FORMATTER, provisioner);
+		// The default importRemover is GJF
+		return create(GJF, provisioner);
 	}
 
 	public static FormatterStep create(String unusedImportRemover, Provisioner provisioner) {
 		Objects.requireNonNull(provisioner, "provisioner");
 		switch (unusedImportRemover) {
-		case DEFAULT_FORMATTER:
+		case GJF:
 			return GoogleJavaFormatStep.createRemoveUnusedImportsOnly(provisioner);
 		case CLEANTHAT:
 			return CleanthatJavaStep.createWithStepName(NAME, CleanthatJavaStep.defaultGroupArtifact(), CleanthatJavaStep.defaultVersion(), "99.9", List.of(CLEANTHAT_MUTATOR), List.of(), false, provisioner);
